Rename 'jasperpath' module to 'paths'
